test bayesian alpha v3

+++ master/scripts/planner/solvers/AlgoPlusCourtsCheminsPourDronesAlphav3doubleDroneFonction.py
@@ -3,7 +3,7 @@
 def test(fi, be, al, it):
     fi = int(fi * 50)
     #be = be * 1
-    al = al * 10
+    al = int(al * 10)
 
     #Lecture
     #with open("plots/bayesian/parameters", 'r') as f:
